# docs/openapi.yaml
openapi: 3.0.0
paths:
  /users/{id}:
    get:
      operationId: UsersController_getProfile
      summary: Получить профиль по id
      description: "Публичные данные User: avatar, bio, isVerified,
        creatorProfile/companyProfile и т.д. Email не возвращается (он хранится
        в Account). isEmailConfirmed виден только владельцу профиля. Доступен
        любому авторизованному пользователю."
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: Данные профиля
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    type: object
                    properties:
                      id:
                        type: string
                        format: uuid
                      role:
                        type: string
                        enum:
                          - CREATOR
                          - COMPANY
                      membershipRole:
                        type: string
                        enum:
                          - OWNER
                          - ADMIN
                          - EDITOR
                          - VIEWER
                      contacts:
                        type: array
                        nullable: true
                        items:
                          type: object
                          properties:
                            type:
                              type: string
                              enum:
                                - phone
                                - telegram
                                - whatsapp
                                - instagram
                                - youtube
                                - vk
                                - website
                                - other
                            value:
                              type: string
                              example: "+79991234567"
                            label:
                              type: string
                              nullable: true
                              example: WhatsApp
                          required:
                            - type
                            - value
                      person:
                        type: object
                        nullable: true
                        properties:
                          height:
                            type: string
                            nullable: true
                            example: "180"
                          weight:
                            type: string
                            nullable: true
                            example: "75"
                          size:
                            type: string
                            nullable: true
                            example: M
                          birthday:
                            type: string
                            nullable: true
                            example: 1995-06-15
                          gender:
                            type: string
                            nullable: true
                            example: male
                          parameters:
                            type: string
                            nullable: true
                            example: 90-60-90
                      phone:
                        type: string
                        nullable: true
                        example: "+79991234567"
                      location:
                        type: string
                        nullable: true
                      avatar:
                        type: string
                        nullable: true
                      bio:
                        type: string
                        nullable: true
                      followers:
                        type: number
                        example: 0
                      isVerified:
                        type: boolean
                        description: Профиль проверен платформой
                      isEmailConfirmed:
                        type: boolean
                        description: Email профиля (User.email) подтверждён
                      aboutMe:
                        type: string
                        nullable: true
                      name:
                        type: string
                      lastName:
                        type: string
                      companyName:
                        type: string
        "404":
          description: Пользователь не найден
      tags: &a1
        - users
      security: &a2
        - access-token: []
  /users/update:
    patch:
      operationId: UsersController_updateProfile
      summary: Обновить активный профиль
      description: "Единственный эндпоинт редактирования профиля. Обновляет **текущий
        активный** User из JWT. Для CREATOR: name, lastName + общие поля. Для
        COMPANY: companyName + общие поля. Общие поля: phone, contacts, person,
        location, avatar, bio, aboutMe, banner. Передайте `null`, чтобы удалить
        nullable поле (phone, bio и т.д.). person мержится с существующим:
        `person: { height: null }` удаляет ключ, `person: null` — весь объект.
        VIEWER получит 403. Для смены профиля перед редактированием —
        switch-profile."
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserDto"
      responses:
        "200":
          description: Обновлённый профиль
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    type: object
                    properties:
                      id:
                        type: string
                        format: uuid
                      role:
                        type: string
                        enum:
                          - CREATOR
                          - COMPANY
                      membershipRole:
                        type: string
                        enum:
                          - OWNER
                          - ADMIN
                          - EDITOR
                          - VIEWER
                      contacts:
                        type: array
                        nullable: true
                        items:
                          type: object
                          properties:
                            type:
                              type: string
                              enum:
                                - phone
                                - telegram
                                - whatsapp
                                - instagram
                                - youtube
                                - vk
                                - website
                                - other
                            value:
                              type: string
                              example: "+79991234567"
                            label:
                              type: string
                              nullable: true
                              example: WhatsApp
                          required:
                            - type
                            - value
                      person:
                        type: object
                        nullable: true
                        properties:
                          height:
                            type: string
                            nullable: true
                            example: "180"
                          weight:
                            type: string
                            nullable: true
                            example: "75"
                          size:
                            type: string
                            nullable: true
                            example: M
                          birthday:
                            type: string
                            nullable: true
                            example: 1995-06-15
                          gender:
                            type: string
                            nullable: true
                            example: male
                          parameters:
                            type: string
                            nullable: true
                            example: 90-60-90
                      phone:
                        type: string
                        nullable: true
                        example: "+79991234567"
                      location:
                        type: string
                        nullable: true
                      avatar:
                        type: string
                        nullable: true
                      bio:
                        type: string
                        nullable: true
                      followers:
                        type: number
                        example: 0
                      isVerified:
                        type: boolean
                        description: Профиль проверен платформой
                      isEmailConfirmed:
                        type: boolean
                        description: Email профиля (User.email) подтверждён
                      aboutMe:
                        type: string
                        nullable: true
                      name:
                        type: string
                      lastName:
                        type: string
                      companyName:
                        type: string
        "403":
          description: Недостаточно прав (роль VIEWER)
      tags: *a1
      security: *a2
...
info:
  title: Project Back API
  description: >-
    # Project Back API


    Backend для marketplace creator ↔ company. Документация для фронтенда и
    AI-ассистентов.


    ## Модель данных


    | Сущность | Что это |

    |----------|---------|

    | **Account** | Login: email + password. Один человек = один Account. |

    | **User** | Публичный профиль: CREATOR или COMPANY. При регистрации
    создаётся один User на Account. |

    | **Membership** | Доступ Account к User: OWNER / ADMIN / EDITOR / VIEWER. |

    | **Invite** | Приглашение другому человеку управлять **существующим**
    профилем. |


    **Активный профиль** — тот, чей `userId` в JWT (`sub`). Переключение: `POST
    /auth/switch-profile`.


    ## Cookies и CORS


    - После login/register/refresh выставляются httpOnly cookies:
    `access-token`, `refresh-token`.

    - Все запросы с `credentials: 'include'`.

    - CORS origin: `CORS_ORIGIN` (обычно http://localhost:3000).


    ## Сценарии


    ### Первый вход

    1. `POST /auth/register/creator` или `/auth/register/company` — создаёт
    Account + первый профиль.

    2. `POST /auth/login` — вход по email/password.


    ### Совместное управление (менеджер)

    1. Владелец: `POST /auth/invites` — email менеджера + `userId`
    **существующего** профиля + роль.

    2. Менеджер: регистрация/логин **своим** email → `POST
    /auth/invites/accept`.

    3. Менеджер: `GET /auth/profiles` → `POST /auth/switch-profile`.


...

...

...


...

...

...


...

...

...

...

...
servers:
...
components:
  securitySchemes:
    cookie:
      type: apiKey
      in: cookie
      name: access-token
  schemas:
    ContactItemDto:
      type: object
      properties:
        type:
          type: string
          enum:
            - phone
            - telegram
            - whatsapp
            - instagram
            - youtube
            - vk
            - website
            - other
          example: phone
          description: "Тип контакта: phone, telegram, instagram и т.д."
        value:
          type: string
          example: "+79991234567"
          description: Номер телефона или URL соцсети
        label:
          type: string
          example: WhatsApp
          description: "Подпись: «Рабочий», «Личный Telegram»"
      required:
        - type
        - value
    PersonDto:
      type: object
      properties:
        height:
          type: string
          example: "180"
          description: Рост
          nullable: true
        weight:
          type: string
          example: "75"
          description: Вес
          nullable: true
        size:
          type: string
          example: M
          description: Размер одежды
          nullable: true
        birthday:
          type: string
          example: 1995-06-15
          description: Дата рождения
          nullable: true
        gender:
          type: string
          example: male
          description: Пол
          nullable: true
        parameters:
          type: string
          example: 90-60-90
          description: Параметры / замеры
          nullable: true
    UpdateCreatorProfileDto:
      type: object
      properties:
        name:
          type: string
          example: John
        lastName:
          type: string
          example: Doe
    UpdateCompanyProfileDto:
      type: object
      properties:
        companyName:
          type: string
          example: Acme Inc.
    UpdateUserDto:
      type: object
      properties:
        name:
          type: string
          example: John
        lastName:
          type: string
          example: Doe
        companyName:
          type: string
          example: Acme Inc.
        contacts:
          nullable: true
          description: "Контакты: телефоны и ссылки на соцсети. Передайте null, чтобы
            удалить"
          type: array
          items:
            $ref: "#/components/schemas/ContactItemDto"
        person:
          nullable: true
          description: "Антропометрия и параметры. Merge с существующим объектом; null на
            поле удаляет ключ; person: null — очистить весь блок"
          allOf:
            - $ref: "#/components/schemas/PersonDto"
        phone:
          type: string
          example: "+79991234567"
          nullable: true
          description: Основной телефон. Передайте null, чтобы удалить
        location:
          type: string
          example: Moscow
          nullable: true
          description: Передайте null, чтобы удалить
        avatar:
          type: string
          example: https://example.com/avatar.png
          nullable: true
          description: Передайте null, чтобы удалить
        bio:
          type: string
          example: Short bio
          nullable: true
          description: Передайте null, чтобы удалить
        aboutMe:
          type: string
          example: More about me
          nullable: true
          description: Передайте null, чтобы удалить
        banner:
          type: string
          example: https://example.com/banner.png
          nullable: true
          description: Передайте null, чтобы удалить
        creatorProfile:
          $ref: "#/components/schemas/UpdateCreatorProfileDto"
        companyProfile:
          $ref: "#/components/schemas/UpdateCompanyProfileDto"
...